Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/svg/animations/discard-on-discard-expected.txt


FAIL Test the behavior of one discard applied on another discard assert_approx_equals: expected 255 +/- 2 but got 0
